+++ client/src/App.css
@@ -61,4 +61,5 @@
   padding-bottom: 2px;
   margin-bottom: 45px;
   height: 47px;
+  text-decoration: none;
 }
