--- client/src/components/Nav/style.css
@@ -1,39 +1,0 @@
-/* #navbar {
-  background-color: black;
-  /* height: 30px; */
-
-/* } */
-
-/* #nav-searchbar {
-  height: 28px;
-  margin-left: auto;
-}
-
-#nav-search-button {
-  margin-left: 5px;
-  margin-right: auto;
-  background-color: #295939;
-  border: none;
-  color: white;
-}
-
-#nav-profile-button {
-  margin-left: auto;
-  margin-right: 10px;
-  width: 100px;
-  background-color: #295939;
-  border: none;
-  color: white;
-}
-
-#nav-map-button {
-  margin-right: auto;
-  width: 100px;
-  background-color: #295939;
-  border: none;
-  color: white;
-} */
-
-.nav-link {
-  text-decoration: none;
-}
