+++ client/src/components/NavCom/style.css
@@ -1,0 +1,7 @@
+.nav-link {
+  text-decoration: none;
+}
+
+.navbar {
+  color: aliceblue;
+}
