+++ client/src/components/auth-nav.js
@@ -2,7 +2,7 @@ import React from "react";
 import AuthenticationButton from "./authentication-button";
 
 const AuthNav = () => (
-  <div className="navbar-nav ml-auto">
+  <div className="">
     <AuthenticationButton />
   </div>
 );
