+++ client/src/pages/Profile/Profile.css
@@ -1,5 +1,5 @@
 .sky {
-  background-image: linear-gradient(rgb(1, 1, 43), rgb(62, 62, 143), rgb(167, 230, 230));
+  background-image: linear-gradient(rgb(33, 33, 41), rgb(62, 62, 143), rgb(167, 230, 230));
   min-height: 100vh;
   min-width: 100%;
 }
@@ -18,17 +18,18 @@
 
 .username {
   font-family: inherit;
-  color: white;
+  color: aliceblue;
   text-align: center;
   margin-top: 3%;
   margin-bottom: 7%;
 }
 
 .list-group-item {
+  margin-top: 5px;
   background-color: inherit !important;
-  color: white !important;
+  color: aliceblue !important;
   text-align: left;
-  font-size: 18px !important;
+  font-size: 16px !important;
   /* margin-left: 5%; */
 }
